Adopt_A_Robot
A refactoring assignment where we learn how to create a game where the player can adopt a robot using our teacher coding as a reference. 
It may not seem to hard to know how to refactor, however, this was once of the first things I learned. It help me to start thinking on a different way to be able to refactor this project.